# Conflicts:
#	lib/token.h
#	test/testsymboldatabase.cpp

+++ tools/donate_cpu_lib.py
@@ -15,10 +15,12 @@ import shlex
 # Version scheme (MAJOR.MINOR.PATCH) should orientate on "Semantic Versioning" https://semver.org/
 # Every change in this script should result in increasing the version number accordingly (exceptions may be cosmetic
 # changes)
-CLIENT_VERSION = "1.3.7"
+CLIENT_VERSION = "1.3.18"
 
 # Timeout for analysis with Cppcheck in seconds
-CPPCHECK_TIMEOUT = 60 * 60
+CPPCHECK_TIMEOUT = 30 * 60
+
+CPPCHECK_REPO_URL = "https://github.com/danmar/cppcheck.git"
 
 # Return code that is used to mark a timed out analysis
 RETURN_CODE_TIMEOUT = -999
@@ -30,47 +32,69 @@ def check_requirements():
         try:
             subprocess.call([app, '--version'])
         except OSError:
-            print(app + ' is required')
+            print("Error: '{}' is required".format(app))
             result = False
+    try:
+        import psutil
+    except ImportError as e:
+        print("Error: {}. Module is required. ".format(e))
+        result = False
     return result
 
 
-def get_cppcheck(cppcheck_path, work_path):
-    print('Get Cppcheck..')
-    for i in range(5):
-        if os.path.exists(cppcheck_path):
-            try:
-                os.chdir(cppcheck_path)
-                try:
-                    subprocess.check_call(['git', 'checkout', '-f', 'main'])
-                except subprocess.CalledProcessError:
-                    subprocess.check_call(['git', 'checkout', '-f', 'master'])
-                    subprocess.check_call(['git', 'pull'])
-                    subprocess.check_call(['git', 'checkout', 'origin/main', '-b', 'main'])
-                subprocess.check_call(['git', 'pull'])
-            except:
-                print('Failed to update Cppcheck sources! Retrying..')
-                time.sleep(10)
-                continue
-        else:
-            try:
-                subprocess.check_call(['git', 'clone', 'https://github.com/danmar/cppcheck.git', cppcheck_path])
-            except:
-                print('Failed to clone, will try again in 10 minutes..')
-                time.sleep(600)
-                continue
-        time.sleep(2)
-        return True
-    if os.path.exists(cppcheck_path):
-        print('Failed to update Cppcheck sources, trying a fresh clone..')
+# Try and retry with exponential backoff if an exception is raised
+def try_retry(fun, fargs=(), max_tries=5):
+    sleep_duration = 5.0
+    for i in range(max_tries):
         try:
-            os.chdir(work_path)
-            shutil.rmtree(cppcheck_path)
-            get_cppcheck(cppcheck_path, work_path)
-        except:
-            print('Failed to remove Cppcheck folder, please manually remove ' + work_path)
-            return False
-    return False
+            return fun(*fargs)
+        except KeyboardInterrupt as e:
+            # Do not retry in case of user abort
+            raise e
+        except BaseException as e:
+            if i < max_tries - 1:
+                print("{} in {}: {}".format(type(e).__name__, fun.__name__, str(e)))
+                print("Trying {} again in {} seconds".format(fun.__name__, sleep_duration))
+                time.sleep(sleep_duration)
+                sleep_duration *= 2.0
+            else:
+                print("Maximum number of tries reached for {}".format(fun.__name__))
+                raise e
+
+
+def clone_cppcheck(repo_path, migrate_from_path):
+    repo_git_dir = os.path.join(repo_path, '.git')
+    if os.path.exists(repo_git_dir):
+        return
+    # Attempt to migrate clone directory used prior to 1.3.17
+    if os.path.exists(migrate_from_path):
+        os.rename(migrate_from_path, repo_path)
+    else:
+        # A shallow git clone (depth = 1) is enough for building and scanning.
+        # Do not checkout until fetch_cppcheck_version.
+        subprocess.check_call(['git', 'clone', '--depth=1', '--no-checkout', CPPCHECK_REPO_URL, repo_path])
+        # Checkout an empty branch to allow "git worktree add" for main later on
+    try:
+        # git >= 2.27
+        subprocess.check_call(['git', 'switch', '--orphan', 'empty'], cwd=repo_path)
+    except subprocess.CalledProcessError:
+        subprocess.check_call(['git', 'checkout','--orphan', 'empty'], cwd=repo_path)
+
+
+def checkout_cppcheck_version(repo_path, version, cppcheck_path):
+    if not os.path.isabs(cppcheck_path):
+        raise ValueError("cppcheck_path is not an absolute path")
+    if os.path.exists(cppcheck_path):
+        subprocess.check_call(['git', 'checkout' , '-f', version], cwd=cppcheck_path)
+        # It is possible to pull branches, not tags
+        if version == 'main':
+            subprocess.check_call(['git', 'pull'], cwd=cppcheck_path)
+    else:
+        if version != 'main':
+            # Since this is a shallow clone, explicitly fetch the remote version tag
+            refspec = 'refs/tags/' + version + ':ref/tags/' + version
+            subprocess.check_call(['git', 'fetch', '--depth=1', 'origin', refspec], cwd=repo_path)
+        subprocess.check_call(['git', 'worktree', 'add', cppcheck_path,  version], cwd=repo_path)
 
 
 def get_cppcheck_info(cppcheck_path):
@@ -81,28 +105,18 @@ def get_cppcheck_info(cppcheck_path):
         return ''
 
 
-def compile_version(work_path, jobs, version):
-    if os.path.isfile(work_path + '/' + version + '/cppcheck'):
+def compile_version(cppcheck_path, jobs):
+    if os.path.isfile(os.path.join(cppcheck_path, 'cppcheck')):
         return True
-    os.chdir(work_path + '/cppcheck')
-    subprocess.call(['git', 'checkout', version])
-    subprocess.call(['make', 'clean'])
-    subprocess.call(['make', jobs, 'MATCHCOMPILER=yes', 'CXXFLAGS=-O2 -g'])
-    if os.path.isfile(work_path + '/cppcheck/cppcheck'):
-        os.mkdir(work_path + '/' + version)
-        destPath = work_path + '/' + version + '/'
-        subprocess.call(['cp', '-R', work_path + '/cppcheck/cfg', destPath])
-        subprocess.call(['cp', 'cppcheck', destPath])
-    subprocess.call(['git', 'checkout', 'main'])
-    try:
-        subprocess.call([work_path + '/' + version + '/cppcheck', '--version'])
-    except OSError:
-        return False
-    return True
+    # Build
+    ret = compile_cppcheck(cppcheck_path, jobs)
+    # Clean intermediate build files
+    subprocess.call(['git', 'clean', '-f', '-d', '-x', '--exclude', 'cppcheck'], cwd=cppcheck_path)
+    return ret
 
 
 def compile_cppcheck(cppcheck_path, jobs):
-    print('Compiling Cppcheck..')
+    print('Compiling {}'.format(os.path.basename(cppcheck_path)))
     try:
         os.chdir(cppcheck_path)
         subprocess.call(['make', jobs, 'MATCHCOMPILER=yes', 'CXXFLAGS=-O2 -g'])
@@ -165,19 +179,19 @@ def handle_remove_readonly(func, path, exc):
         func(path)
 
 
-def remove_tree(folderName):
-    if not os.path.exists(folderName):
+def remove_tree(folder_name):
+    if not os.path.exists(folder_name):
         return
     count = 5
     while count > 0:
         count -= 1
         try:
-            shutil.rmtree(folderName, onerror=handle_remove_readonly)
+            shutil.rmtree(folder_name, onerror=handle_remove_readonly)
             break
         except OSError as err:
             time.sleep(30)
             if count == 0:
-                print('Failed to cleanup {}: {}'.format(folderName, err))
+                print('Failed to cleanup {}: {}'.format(folder_name, err))
                 sys.exit(1)
 
 
@@ -214,7 +228,6 @@ def unpack_package(work_path, tgz):
     temp_path = work_path + '/temp'
     remove_tree(temp_path)
     os.mkdir(temp_path)
-    os.chdir(temp_path)
     found = False
     if tarfile.is_tarfile(tgz):
         with tarfile.open(tgz) as tf:
@@ -223,16 +236,14 @@ def unpack_package(work_path, tgz):
                     # Skip dangerous file names
                     continue
                 elif member.name.lower().endswith(('.c', '.cpp', '.cxx', '.cc', '.c++', '.h', '.hpp',
-                                                   '.h++', '.hxx', '.hh', '.tpp', '.txx', '.qml',
-                                                   '.sln', '.vcproj', '.vcxproj')):
+                                                   '.h++', '.hxx', '.hh', '.tpp', '.txx', '.ipp', '.ixx', '.qml')):
                     try:
-                        tf.extract(member.name)
+                        tf.extract(member.name, temp_path)
                         found = True
                     except OSError:
                         pass
                     except AttributeError:
                         pass
-    os.chdir(work_path)
     return found
 
 
@@ -254,23 +265,39 @@ def has_include(path, includes):
 
 def run_command(cmd):
     print(cmd)
-    startTime = time.time()
+    start_time = time.time()
+    comm = None
     p = subprocess.Popen(shlex.split(cmd), stdout=subprocess.PIPE, stderr=subprocess.PIPE, preexec_fn=os.setsid)
     try:
         comm = p.communicate(timeout=CPPCHECK_TIMEOUT)
         return_code = p.returncode
+        p = None
     except subprocess.TimeoutExpired:
-        os.killpg(os.getpgid(p.pid), signal.SIGTERM)  # Send the signal to all the process groups
-        comm = p.communicate()
+        import psutil
         return_code = RETURN_CODE_TIMEOUT
+        # terminate all the child processes so we get messages about which files were hanging
+        child_procs = psutil.Process(p.pid).children(recursive=True)
+        if len(child_procs) > 0:
+            for child in child_procs:
+                child.terminate()
+            try:
+                # call with timeout since it might get stuck e.g. gcc-arm-none-eabi
+                comm = p.communicate(timeout=5)
+                p = None
+            except subprocess.TimeoutExpired:
+                pass
+    finally:
+        if p:
+            os.killpg(os.getpgid(p.pid), signal.SIGTERM)  # Send the signal to all the process groups
+            comm = p.communicate()
     stop_time = time.time()
     stdout = comm[0].decode(encoding='utf-8', errors='ignore')
     stderr = comm[1].decode(encoding='utf-8', errors='ignore')
-    elapsed_time = stop_time - startTime
+    elapsed_time = stop_time - start_time
     return return_code, stdout, stderr, elapsed_time
 
 
-def scan_package(work_path, cppcheck_path, jobs, libraries):
+def scan_package(work_path, cppcheck_path, jobs, libraries, capture_callstack = True):
     print('Analyze..')
     os.chdir(work_path)
     libs = ''
@@ -278,65 +305,20 @@ def scan_package(work_path, cppcheck_path, jobs, libraries):
         if os.path.exists(os.path.join(cppcheck_path, 'cfg', library + '.cfg')):
             libs += '--library=' + library + ' '
 
+    dir_to_scan = 'temp'
+
     # Reference for GNU C: https://gcc.gnu.org/onlinedocs/cpp/Common-Predefined-Macros.html
-    options = libs + jobs + ' --showtime=top5 --check-library --inconclusive --enable=style,information --template=daca2 -rp=temp'
-    if os.path.isfile('temp/tortoisesvn/TortoiseSVN.sln'):
-        options = options.replace('--library=posix ', '')
-        options = options.replace('--library=gnu ', '')
-        options = '--library=windows ' + options
-        options += ' --platform=win64 --project=temp/tortoisesvn/TortoiseSVN.sln'
-    else:
-        options += ' -D__GNUC__ --platform=unix64 temp'
+    options = libs + ' --showtime=top5 --check-library --inconclusive --enable=style,information --template=daca2'
+    options += ' -D__GNUC__ --platform=unix64'
+    options += ' -rp={}'.format(dir_to_scan)
     cppcheck_cmd = cppcheck_path + '/cppcheck' + ' ' + options
-    cmd = 'nice ' + cppcheck_cmd
+    cmd = 'nice ' + cppcheck_cmd + ' ' + jobs + ' ' + dir_to_scan
     returncode, stdout, stderr, elapsed_time = run_command(cmd)
-    sig_num = -1
-    sig_msg = 'Internal error: Child process crashed with signal '
-    sig_pos = stderr.find(sig_msg)
-    if sig_pos != -1:
-        sig_start_pos = sig_pos + len(sig_msg)
-        sig_num = int(stderr[sig_start_pos:stderr.find(' ', sig_start_pos)])
-    print('cppcheck finished with ' + str(returncode) + ('' if sig_num == -1 else ' (signal ' + str(sig_num) + ')'))
-    if returncode == RETURN_CODE_TIMEOUT:
-        print('Timeout!')
-        return returncode, stdout, '', elapsed_time, options, ''
-    # generate stack trace for SIGSEGV, SIGABRT, SIGILL, SIGFPE, SIGBUS
-    if returncode in (-11, -6, -4, -8, -7) or sig_num in (11, 6, 4, 8, 7):
-        print('Crash!')
-        stacktrace = ''
-        if cppcheck_path == 'cppcheck':
-            # re-run within gdb to get a stacktrace
-            cmd = 'gdb --batch --eval-command=run --eval-command="bt 50" --return-child-result --args ' + cppcheck_cmd + " -j1"
-            dummy, st_stdout, dummy, dummy = run_command(cmd)
-            gdb_pos = st_stdout.find(" received signal")
-            if not gdb_pos == -1:
-                last_check_pos = st_stdout.rfind('Checking ', 0, gdb_pos)
-                if last_check_pos == -1:
-                    stacktrace = st_stdout[gdb_pos:]
-                else:
-                    stacktrace = st_stdout[last_check_pos:]
-        # if no stacktrace was generated return the original stdout
-        if not stacktrace:
-            stacktrace = stdout
-        return returncode, stacktrace, '', returncode, options, ''
-    if returncode != 0:
-        print('Error!')
-        if returncode > 0:
-            returncode = -100-returncode
-        return returncode, stdout, '', returncode, options, ''
-    err_s = 'Internal error: Child process crashed with signal '
-    err_pos = stderr.find(err_s)
-    if err_pos != -1:
-        print('Error!')
-        pos2 = stderr.find(' [cppcheckError]', err_pos)
-        signr = int(stderr[err_pos+len(err_s):pos2])
-        return -signr, '', '', -signr, options, ''
-    thr_pos = stderr.find('#### ThreadExecutor')
-    if thr_pos != -1:
-        print('Thread!')
-        return -222, stderr[thr_pos:], '', -222, options, ''
+
+    # collect messages
     information_messages_list = []
     issue_messages_list = []
+    internal_error_messages_list = []
     count = 0
     for line in stderr.split('\n'):
         if ': information: ' in line:
@@ -345,6 +327,8 @@ def scan_package(work_path, cppcheck_path, jobs, libraries):
             issue_messages_list.append(line + '\n')
             if re.match(r'.*:[0-9]+:.*\]$', line):
                 count += 1
+            if ': error: Internal error: ' in line:
+                internal_error_messages_list.append(line + '\n')
     print('Number of issues: ' + str(count))
     # Collect timing information
     stdout_lines = stdout.split('\n')
@@ -361,6 +345,75 @@ def scan_package(work_path, cppcheck_path, jobs, libraries):
             timing_info_list.insert(0, ' ' + reverse_line + '\n')
             current_timing_lines += 1
     timing_str = ''.join(timing_info_list)
+
+    # detect errors
+    sig_file = None
+    sig_num = -1
+    for ie_line in internal_error_messages_list:
+        # temp/dlib-19.10/dlib/test/dnn.cpp:0:0: error: Internal error: Child process crashed with signal 11 [cppcheckError]
+        if 'Child process crashed with signal' in ie_line:
+            sig_file = ie_line.split(':')[0]
+            sig_msg = 'signal '
+            sig_pos = ie_line.find(sig_msg)
+            if sig_pos != -1:
+                sig_start_pos = sig_pos + len(sig_msg)
+                sig_num = int(ie_line[sig_start_pos:ie_line.find(' ', sig_start_pos)])
+            # break on the first signalled file for now
+            break
+    print('cppcheck finished with ' + str(returncode) + ('' if sig_num == -1 else ' (signal ' + str(sig_num) + ')'))
+
+    if returncode == RETURN_CODE_TIMEOUT:
+        print('Timeout!')
+        return returncode, ''.join(internal_error_messages_list), '', elapsed_time, options, ''
+
+    # generate stack trace for SIGSEGV, SIGABRT, SIGILL, SIGFPE, SIGBUS
+    has_error = returncode in (-11, -6, -4, -8, -7)
+    has_sig = sig_num in (11, 6, 4, 8, 7)
+    if has_error or has_sig:
+        print('Crash!')
+        # make sure we have the actual error code set
+        if has_sig:
+            returncode = -sig_num
+        stacktrace = ''
+        if capture_callstack:
+            # re-run within gdb to get a stacktrace
+            cmd = 'gdb --batch --eval-command=run --eval-command="bt 50" --return-child-result --args ' + cppcheck_cmd + " -j1 "
+            if sig_file is not None:
+                cmd += sig_file
+            else:
+                cmd += dir_to_scan
+            _, st_stdout, _, _ = run_command(cmd)
+            gdb_pos = st_stdout.find(" received signal")
+            if not gdb_pos == -1:
+                last_check_pos = st_stdout.rfind('Checking ', 0, gdb_pos)
+                if last_check_pos == -1:
+                    stacktrace = st_stdout[gdb_pos:]
+                else:
+                    stacktrace = st_stdout[last_check_pos:]
+        # if no stacktrace was generated return the original stdout or internal errors list
+        if not stacktrace:
+            if has_sig:
+                stacktrace = ''.join(internal_error_messages_list)
+            else:
+                stacktrace = stdout
+        return returncode, stacktrace, '', returncode, options, ''
+
+    if returncode != 0:
+        # returncode is always 1 when this message is written
+        thr_pos = stderr.find('#### ThreadExecutor')
+        if thr_pos != -1:
+            print('Thread!')
+            return -222, stderr[thr_pos:], '', -222, options, ''
+
+        print('Error!')
+        if returncode > 0:
+            returncode = -100-returncode
+        return returncode, stdout, '', returncode, options, ''
+
+    if sig_num != -1:
+        print('Signal!')
+        return -sig_num, ''.join(internal_error_messages_list), '', -sig_num, options, ''
+
     return count, ''.join(issue_messages_list), ''.join(information_messages_list), elapsed_time, options, timing_str
 
 
@@ -379,7 +432,7 @@ def split_results(results):
     return ret
 
 
-def diff_results(work_path, ver1, results1, ver2, results2):
+def diff_results(ver1, results1, ver2, results2):
     print('Diff results..')
     ret = ''
     r1 = sorted(split_results(results1))
@@ -458,11 +511,11 @@ def upload_info(package, info_output, server_address):
 def get_libraries():
     libraries = ['posix', 'gnu']
     library_includes = {'boost': ['<boost/'],
-                       'bsd': ['<sys/queue.h>','<sys/tree.h>','<bsd/','<fts.h>','<db.h>','<err.h>','<vis.h>'],
+                       'bsd': ['<sys/queue.h>', '<sys/tree.h>', '<bsd/', '<fts.h>', '<db.h>', '<err.h>', '<vis.h>'],
                        'cairo': ['<cairo.h>'],
                        'cppunit': ['<cppunit/'],
                        'icu': ['<unicode/', '"unicode/'],
-                       'ginac': ['<ginac/','"ginac/'],
+                       'ginac': ['<ginac/', '"ginac/'],
                        'googletest': ['<gtest/gtest.h>'],
                        'gtk': ['<gtk', '<glib.h>', '<glib-', '<glib/', '<gnome'],
                        'kde': ['<KGlobal>', '<KApplication>', '<KDE/'],
@@ -495,6 +548,11 @@ def get_libraries():
         if has_include('temp', includes):
             libraries.append(library)
     return libraries
+
+
+def get_compiler_version():
+    _, stdout, _, _ = run_command('g++ --version')
+    return stdout.split('\n')[0]
 
 
 my_script_name = os.path.splitext(os.path.basename(sys.argv[0]))[0]
